Create DelegateWeakEventManager.shared.cs
