+++ XamarinCommunityToolkit/Xamarin.CommunityToolkit.csproj
@@ -22,7 +22,9 @@
     <Copyright>© Microsoft Corporation. All rights reserved.</Copyright>
     <RepositoryUrl>https://github.com/xamarin/XamarinCommunityToolkit</RepositoryUrl>
     <PackageReleaseNotes>
+      
       <!-- See if we can host these on MS docs? -->
+    
     </PackageReleaseNotes>
     <DefineConstants>$(DefineConstants);</DefineConstants>
     <UseFullSemVerForNuGet>false</UseFullSemVerForNuGet>
@@ -195,5 +197,10 @@
   <ItemGroup>
     <Compile Remove="build\*.cs" />
     <None Include="build\**\*.cs;build\**\*.targets" Pack="true" PackagePath="build" />
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Update="Helpers\DelegateWeakEventManager.shared.cs">
+      <SubType></SubType>
+    </Compile>
   </ItemGroup>
 </Project>
